Replace LandOwnershipNotification with DismissibleNotification
Now a re-usable component
